Move domain/Entities/Registers -> domain/Entities
